+++ src/TestTrackingDiagrams.Extensions.MongoDB/MongoDbTrackingVerbosity.cs
@@ -1,0 +1,8 @@
+namespace TestTrackingDiagrams.Extensions.MongoDB;
+
+public enum MongoDbTrackingVerbosity
+{
+    Raw,
+    Detailed,
+    Summarised
+}